+++ app/models/base_slot.rb
@@ -56,8 +56,8 @@ class BaseSlot < ActiveRecord::Base
     media_items.image.order(:position)
   end
 
-  def voices
-    media_items.voice.order(:position)
+  def audios
+    media_items.audio.order(:position)
   end
 
   def videos
@@ -193,10 +193,10 @@ class BaseSlot < ActiveRecord::Base
       when 'photo'
         add_media(item.permit(:public_id, :position, :local_id)
                    .merge(media_type: 'image'))
-      when 'voice'
+      when 'audio'
         add_media(
           item.permit(:public_id, :position, :local_id, :duration, :title)
-          .merge(media_type: 'voice'))
+          .merge(media_type: 'audio'))
       when 'video'
         add_media(
           item.permit(:public_id, :position, :local_id, :duration, :thumbnail)