+++ app/models/media_item.rb
@@ -4,12 +4,12 @@ class MediaItem < ActiveRecord::Base
   belongs_to :mediable, polymorphic: true
 
   scope :image, -> { where media_type: 'image' }
-  scope :voice, -> { where media_type: 'voice' }
+  scope :audio, -> { where media_type: 'audio' }
   scope :video, -> { where media_type: 'video' }
 
   validates :media_type,
             presence: true,
-            inclusion: { in: %w(image voice video) }
+            inclusion: { in: %w(image audio video) }
   validates :public_id, presence: true
   validates :position, presence: true, if: :belongs_to_slot?
   validates :mediable_id, presence: true