+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## latest
 - add title to media items of type voice/audio
 - change type of media items duration from int to float
+- rename voice to audio
+
 
 ## v0.5.3 on t2s2-dev
 - increase length of media item local_id from 64 to 512